requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
transmissionrpc==0.11
pydantic==1.10.14
pyjwt==2.8.0
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
gunicorn==23.0.0